App Impersonator Project
In this project I simulated a simple HRV Monitor app in which I have created tables with data that would accurately represent the type of data that could be found in this type of app. I then utilized the UPDATE, DELETE, ALTER, and INSERT INTO commands to replicate the kind of work flow and data manipulation that might occur during the app's operation.